Parameter tuning, Min_Threads added
Set chess piece values to those from the GH compile of Fire 2.2.
Introduce Min_Threads and remove doubling of the CPU number to get the
number of threads. Now it is still possible to have more threads than
processors if one wants to, but such a setting is no longer the default,
which was bad.